+++ doc/releases.rst
@@ -8,6 +8,11 @@ Latest Development Version
 | Version: 1.22.0
 | Released: 2021-11-10
 
+- with :ref:`value command <value command>`, only the first few characters of
+  the field name need be specified.
+- added :ref:`additional_account_templates <additional_account_templates 
+  setting>` setting.
+
 
 1.22 (2021-11-10)
 -----------------